added some Type and ID tests in the backend, sorted the json list, changed POST request HTTP status from 403 to 409 when a resource already exists

+++ backend/rest_test.go
@@ -16,12 +16,6 @@ import (
 	"github.com/restic/restic/backend/rest"
 )
 
-func setupRestBackend(t *testing.T) *rest.Rest {
-	url, _ := url.Parse("http://localhost:8000")
-	backend, _ := rest.Open(url)
-	return backend
-}
-
 func TestRestBackend(t *testing.T) {
 
 	// Initializing a temporary direcory for the rest backend.
@@ -30,45 +24,43 @@ func TestRestBackend(t *testing.T) {
 
 	dirs := []string{
 		path,
-		filepath.Join(path, "data"),
-		filepath.Join(path, "snapshot"),
-		filepath.Join(path, "index"),
-		filepath.Join(path, "lock"),
-		filepath.Join(path, "key"),
-		filepath.Join(path, "temp"),
+		filepath.Join(path, string(backend.Data)),
+		filepath.Join(path, string(backend.Snapshot)),
+		filepath.Join(path, string(backend.Index)),
+		filepath.Join(path, string(backend.Lock)),
+		filepath.Join(path, string(backend.Key)),
 	}
 
 	for _, d := range dirs {
 		os.MkdirAll(d, backend.Modes.Dir)
 	}
 
-	// Initialize the router for the repository requests.
 	r := mux.NewRouter()
 
-	// Check if the repository has already been initialized.
+	// Check if a configuration exists.
 	r.HandleFunc("/config", func(w http.ResponseWriter, r *http.Request) {
 		file := filepath.Join(path, "config")
 		if _, err := os.Stat(file); err != nil {
-			http.Error(w, "Repository not found", 404)
+			http.Error(w, "404 repository not found", 404)
 		}
 	}).Methods("HEAD")
 
-	// Get the configuration of the repository.
+	// Get the configuration.
 	r.HandleFunc("/config", func(w http.ResponseWriter, r *http.Request) {
 		file := filepath.Join(path, "config")
 		if _, err := os.Stat(file); err == nil {
 			bytes, _ := ioutil.ReadFile(file)
 			w.Write(bytes)
 		} else {
-			http.Error(w, "Repository not found", 404)
+			http.Error(w, "404 repository not found", 404)
 		}
 	}).Methods("GET")
 
-	// Initialize the repository and save the configuration.
+	// Save the configuration.
 	r.HandleFunc("/config", func(w http.ResponseWriter, r *http.Request) {
 		file := filepath.Join(path, "config")
 		if _, err := os.Stat(file); err == nil {
-			http.Error(w, "Repository already initialized", 403)
+			http.Error(w, "409 repository already initialized", 409)
 		} else {
 			bytes, _ := ioutil.ReadAll(r.Body)
 			ioutil.WriteFile(file, bytes, 0600)
@@ -79,8 +71,12 @@ func TestRestBackend(t *testing.T) {
 	// List the blobs of a given type.
 	r.HandleFunc("/{type}/", func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
-		blobType := filepath.Clean(vars["type"])
-		path := filepath.Join(path, blobType)
+		blobType, errType := backend.ParseType(filepath.Clean(vars["type"]))
+		if errType != nil {
+			http.Error(w, "403 invalid blob type", 403)
+			return
+		}
+		path := filepath.Join(path, string(blobType))
 		files, _ := ioutil.ReadDir(path)
 		names := make([]string, len(files))
 		for i, f := range files {
@@ -93,35 +89,59 @@ func TestRestBackend(t *testing.T) {
 	// Check if a blob of a given type exists.
 	r.HandleFunc("/{type}/{blob}", func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
-		blobType := filepath.Clean(vars["type"])
-		blobID := filepath.Clean(vars["blob"])
-		blob := filepath.Join(path, blobType, blobID)
+		blobType, errType := backend.ParseType(filepath.Clean(vars["type"]))
+		if errType != nil {
+			http.Error(w, "403 invalid blob type", 403)
+			return
+		}
+		blobID, errID := backend.ParseID(vars["blob"])
+		if errID != nil {
+			http.Error(w, "403 invalid blob ID", 403)
+			return
+		}
+		blob := filepath.Join(path, string(blobType), blobID.String())
 		if _, err := os.Stat(blob); err != nil {
-			http.Error(w, "Blob not found", 404)
+			http.Error(w, "404 blob not found", 404)
 		}
 	}).Methods("HEAD")
 
 	// Get a blob of a given type.
 	r.HandleFunc("/{type}/{blob}", func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
-		blobType := filepath.Clean(vars["type"])
-		blobID := filepath.Clean(vars["blob"])
-		blob := filepath.Join(path, blobType, blobID)
+		blobType, errType := backend.ParseType(filepath.Clean(vars["type"]))
+		if errType != nil {
+			http.Error(w, "403 invalid blob type", 403)
+			return
+		}
+		blobID, errID := backend.ParseID(vars["blob"])
+		if errID != nil {
+			http.Error(w, "403 invalid blob ID", 403)
+			return
+		}
+		blob := filepath.Join(path, string(blobType), blobID.String())
 		if file, err := os.Open(blob); err == nil {
 			http.ServeContent(w, r, "", time.Unix(0, 0), file)
 		} else {
-			http.Error(w, "Blob not found", 404)
+			http.Error(w, "404 blob not found", 404)
 		}
 	}).Methods("GET")
 
 	// Save a blob of a given type.
 	r.HandleFunc("/{type}/{blob}", func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
-		blobType := filepath.Clean(vars["type"])
-		blobID := filepath.Clean(vars["blob"])
-		blob := filepath.Join(path, blobType, blobID)
+		blobType, errType := backend.ParseType(filepath.Clean(vars["type"]))
+		if errType != nil {
+			http.Error(w, "403 invalid blob type", 403)
+			return
+		}
+		blobID, errID := backend.ParseID(vars["blob"])
+		if errID != nil {
+			http.Error(w, "403 invalid blob ID", 403)
+			return
+		}
+		blob := filepath.Join(path, string(blobType), blobID.String())
 		if _, err := os.Stat(blob); err == nil {
-			http.Error(w, "Blob already uploaded", 403)
+			http.Error(w, "409 blob already uploaded", 409)
 		} else {
 			bytes, _ := ioutil.ReadAll(r.Body)
 			ioutil.WriteFile(blob, bytes, 0600)
@@ -131,13 +151,21 @@ func TestRestBackend(t *testing.T) {
 	// Delete a blob of a given type.
 	r.HandleFunc("/{type}/{blob}", func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
-		blobType := filepath.Clean(vars["type"])
-		blobID := filepath.Clean(vars["blob"])
-		blob := filepath.Join(path, blobType, blobID)
+		blobType, errType := backend.ParseType(filepath.Clean(vars["type"]))
+		if errType != nil {
+			http.Error(w, "403 invalid blob type", 403)
+			return
+		}
+		blobID, errID := backend.ParseID(vars["blob"])
+		if errID != nil {
+			http.Error(w, "403 invalid blob ID", 403)
+			return
+		}
+		blob := filepath.Join(path, string(blobType), blobID.String())
 		if _, err := os.Stat(blob); err == nil {
 			os.Remove(blob)
 		} else {
-			http.Error(w, "Blob not found", 404)
+			http.Error(w, "404 blob not found", 404)
 		}
 	}).Methods("DELETE")
